MLXPK-1847 Update dataset factory to use registry
* moved NSSDataset from core into usecases/nss
