+++ src/ng_model_gym/usecases/nss/configs/default.json
@@ -4,6 +4,8 @@
         "version": "1"
     },
     "dataset": {
+        "name": "NSS",
+        "version": "1",
         "path": {
             "train": null,
             "validation": null,
