+++ tests/usecases/nss/unit/test_nss_dataset.py
@@ -8,8 +8,8 @@ import unittest
 import torch
 from safetensors import safe_open
 
-from ng_model_gym.core.data.dataloader import NSSDataset
 from ng_model_gym.core.data.utils import DataLoaderMode, DatasetType
+from ng_model_gym.usecases.nss.data.dataset import NSSDataset
 from tests.testing_utils import create_simple_params
